+++ src/main/java/erel/gui/MainWindow.java
@@ -1,6 +1,6 @@
 package erel.gui;
 
-import erel.Erel;
+import erel.ui.Erel;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
 import javafx.scene.control.ScrollPane;
